feat: (#22) Получение списка свойств глобального контекста + тесты
