+++ src/test/java/org/silverbulleters/bsl/platform/context/BSLPlatformContextTest.java
@@ -26,8 +26,6 @@ import org.silverbulleters.bsl.platform.context.internal.PlatformContextStorage;
 import org.silverbulleters.bsl.platform.context.internal.util.ReadDataCollector;
 import org.silverbulleters.bsl.platform.context.platform.PlatformEdition;
 import org.silverbulleters.bsl.platform.context.types.PlatformTypeIdentifier;
-import org.silverbulleters.bsl.platform.context.types.PrimitiveType;
-import org.silverbulleters.bsl.platform.context.types.Resource;
 
 import java.util.Arrays;
 import java.util.List;
@@ -76,17 +74,6 @@ class BSLPlatformContextTest {
     Arrays.stream(PlatformEdition.values()).forEach(this::checkDataVersion);
   }
 
-  @Test
-  void testGetTypeByName() {
-    var edition = PlatformEdition.VERSION_8_3_10;
-    var context = new BSLPlatformContext(List.of(edition));
-    var names = new Resource("Запрос", "Query");
-    checkGetTypeByName(context, edition, names);
-
-    var type = context.getTypeByName(edition, "ЧтотоДругое");
-    assertThat(type).isEqualTo(PrimitiveType.UNKNOWN_TYPE);
-  }
-
   private void checkDataVersion(PlatformEdition platformEdition) {
     var contextStorage = new PlatformContextStorage();
     var data = ReadDataCollector.readToPlatformContext(platformEdition, contextStorage.getTypeRefs());
@@ -108,13 +95,4 @@ class BSLPlatformContextTest {
     });
   }
 
-  private static void checkGetTypeByName(BSLPlatformContext context, PlatformEdition edition, Resource names) {
-    var type = context.getTypeByName(edition, names.getNameRu());
-    assertThat(type).isNotEqualTo(PrimitiveType.UNKNOWN_TYPE);
-    assertThat(type.getName()).isEqualTo(names);
-
-    type = context.getTypeByName(edition, names.getNameEn());
-    assertThat(type).isNotEqualTo(PrimitiveType.UNKNOWN_TYPE);
-    assertThat(type.getName()).isEqualTo(names);
-  }
 }
